Tightening up AWS SDK dependencies.
The original pom.xml had a dependency on the top-level AWS SDK
which meant that it pulled in ALL of the AWS SDK dependencies.
A perusal of the code shows there are actually only dependencies
on the cloudwatch, dynamodb, and kinesis AWS SDKs.

This update changes the POM dependencies to depend upon the
AWS SDK's cloudwatch, dynamodb, and kinesis modules instead
of the high-level AWS SDK.